+++ src/main/java/com/eskcti/algafoodapi/api/exceptionhandler/ApiExceptionHandler.java
@@ -1,0 +1,44 @@
+package com.eskcti.algafoodapi.api.exceptionhandler;
+
+import com.eskcti.algafoodapi.domain.exceptions.BusinessException;
+import com.eskcti.algafoodapi.domain.exceptions.EntityNotFoundException;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.HttpMediaTypeNotSupportedException;
+import org.springframework.web.bind.annotation.ControllerAdvice;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+
+import java.time.LocalDateTime;
+
+@ControllerAdvice
+public class ApiExceptionHandler {
+    @ExceptionHandler(EntityNotFoundException.class)
+    public ResponseEntity<?> handleEntityNotFoundException(EntityNotFoundException e) {
+        Problem problem = Problem.builder()
+                .dateTime(LocalDateTime.now())
+                .message(e.getMessage())
+                .build();
+        return ResponseEntity.status(HttpStatus.NOT_FOUND)
+                .body(problem);
+    }
+
+    @ExceptionHandler(BusinessException.class)
+    public ResponseEntity<?> handleBusinessException(BusinessException e) {
+        Problem problem = Problem.builder()
+                .dateTime(LocalDateTime.now())
+                .message(e.getMessage())
+                .build();
+        return ResponseEntity.status(HttpStatus.BAD_REQUEST)
+                .body(problem);
+    }
+
+    @ExceptionHandler(HttpMediaTypeNotSupportedException.class)
+    public ResponseEntity<?> handleHttpMediaTypeNotSupportedException(HttpMediaTypeNotSupportedException e) {
+        Problem problem = Problem.builder()
+                .dateTime(LocalDateTime.now())
+                .message(e.getMessage())
+                .build();
+        return ResponseEntity.status(HttpStatus.UNSUPPORTED_MEDIA_TYPE)
+                .body(problem);
+    }
+}
