+++ src/main/java/com/eskcti/algafoodapi/api/resources/CityController.java
@@ -58,24 +58,4 @@ public class CityController {
     public void delete(@PathVariable Long id) {
         cityService.remove(id);
     }
-
-    @ExceptionHandler(EntityNotFoundException.class)
-    public ResponseEntity<?> handleEntityNotFoundException(EntityNotFoundException e) {
-        Problem problem = Problem.builder()
-                .dateTime(LocalDateTime.now())
-                .message(e.getMessage())
-                .build();
-        return ResponseEntity.status(HttpStatus.NOT_FOUND)
-                .body(problem);
-    }
-
-    @ExceptionHandler(BusinessException.class)
-    public ResponseEntity<?> handleBusinessException(BusinessException e) {
-        Problem problem = Problem.builder()
-                .dateTime(LocalDateTime.now())
-                .message(e.getMessage())
-                .build();
-        return ResponseEntity.status(HttpStatus.BAD_REQUEST)
-                .body(problem);
-    }
 }
